Remove pycache and ignore generated files
